+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ```
     * Removed the updateOrder reques on SDK pre-payment event.
     * Changes into the logic who check for plugin updates.
+    * Check if Order total and currency are correct when DMN come.
 ```
 
 # 1.7